install sweetalert2 dan pengaplikasian alert crud

+++ resources/views/kriteria/script.blade.php
@@ -64,16 +64,37 @@
                 var deleteUrl = form.attr('action');
                 var currentPage = tableKriteria.page();
 
-                $.ajax({
-                    url: deleteUrl,
-                    type: 'DELETE',
-                    success: function(response) {
-                        tableKriteria.ajax.reload();
-                        tableKriteria.page(currentPage).draw('page');
-                        console.log(response.message)
-                    },
-                    error: function(xhr) {
-                        console.log(xhr.responseJSON.message);
+                Swal.fire({
+                    title: 'Apakah Anda yakin?',
+                    text: 'Data yang dihapus tidak bisa dikembalikan!',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#d33',
+                    cancelButtonColor: '#3085d6',
+                    confirmButtonText: 'Ya, hapus!',
+                    cancelButtonText: 'Batal'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $.ajax({
+                            url: deleteUrl,
+                            type: 'DELETE',
+                            success: function(response) {
+                                tableKriteria.ajax.reload();
+                                tableKriteria.page(currentPage).draw('page');
+                                Swal.fire({
+                                    title: "Success",
+                                    text: response.message,
+                                    icon: "success"
+                                });
+                            },
+                            error: function(xhr) {
+                                Swal.fire({
+                                    title: "Error",
+                                    text: xhr.responseJSON.message,
+                                    icon: "error"
+                                });
+                            }
+                        });
                     }
                 });
             });
@@ -94,10 +115,18 @@
                     $('#dataKriteria').DataTable().ajax.reload();
                     $('#dataKriteria').DataTable().page(currentPage).draw('page');
                     $('#kriteriaModal').modal('hide');
-                    console.log(response.message);
+                    Swal.fire({
+                        title: "Success",
+                        text: response.message,
+                        icon: "success"
+                    });
                 },
                 error: function(xhr) {
-                    console.log(xhr.responseJSON.message);
+                    Swal.fire({
+                        title: "Error",
+                        text: xhr.responseJSON.message,
+                        icon: "error"
+                    });
                 }
             });
         });
