+++ resources/views/sekolah/script.blade.php
@@ -67,19 +67,41 @@
                 var form = $(this).closest('form');
                 var deleteUrl = form.attr('action');
                 var currentPage = tableSekolah.page();
-                
-                $.ajax({
-                    url: deleteUrl,
-                    type: 'DELETE',
-                    success: function(response) {
-                        tableSekolah.ajax.reload();
-                        tableSekolah.page(currentPage).draw('page');
-                        console.log(response.message);
-                    },
-                    error: function(xhr) {
-                        console.log(xhr.responseJSON.message);
+
+                Swal.fire({
+                    title: 'Apakah Anda yakin?',
+                    text: 'Data yang dihapus tidak bisa dikembalikan!',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#d33',
+                    cancelButtonColor: '#3085d6',
+                    confirmButtonText: 'Ya, hapus!',
+                    cancelButtonText: 'Batal'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $.ajax({
+                            url: deleteUrl,
+                            type: 'DELETE',
+                            success: function(response) {
+                                tableSekolah.ajax.reload();
+                                $('#dataKriteriaSekolah').DataTable().ajax.reload();
+                                tableSekolah.page(currentPage).draw('page');
+                                Swal.fire({
+                                    title: 'Success',
+                                    text: response.message,
+                                    icon: 'success',
+                                });
+                            },
+                            error: function(xhr) {
+                                Swal.fire({
+                                    title: 'Error',
+                                    text: xhr.responseJSON.message,
+                                    icon: 'error',
+                                });
+                            }
+                        });
                     }
-                })
+                });
             });
 
             $('#wilayah_kecamatan').on('change', function () {
@@ -122,10 +144,18 @@
                     $('#dataSekolah').DataTable().ajax.reload();
                     $('#dataSekolah').DataTable().page(currentPage).draw('page');
                     $('#sekolahModal').modal('hide');
-                    console.log(response.message);
+                    Swal.fire({
+                        title: 'Success',
+                        text: response.message,
+                        icon: 'success',
+                    });
                 },
                 error: function(xhr) {
-                    console.log(xhr.responseJSON.message);
+                    Swal.fire({
+                        title: 'Error',
+                        text: xhr.responseJSON.message,
+                        icon: 'error',
+                    });
                 }
             });
         });
