+++ resources/views/wilayah/kelurahan/script.blade.php
@@ -54,19 +54,40 @@
                 var deleteUrl = form.attr('action');
                 var currentPage = tableKelurahan.page();
 
-                $.ajax({
-                    url: deleteUrl,
-                    type: 'DELETE',
-                    success: function(response) {
-                        tableKelurahan.ajax.reload();
-                        // tableNilaiKriteiaWilayah.ajax.reload();
-                        tableKelurahan.page(currentPage).draw('page');
-                        console.log(response.message);
-                    },
-                    error: function(xhr) {
-                        console.log(xhr.responseJSON.message);
+                Swal.fire({
+                    title: 'Apakah Anda yakin?',
+                    text: 'Data yang dihapus tidak bisa dikembalikan!',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#d33',
+                    cancelButtonColor: '#3085d6',
+                    confirmButtonText: 'Ya, hapus!',
+                    cancelButtonText: 'Batal'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $.ajax({
+                            url: deleteUrl,
+                            type: 'DELETE',
+                            success: function(response) {
+                                tableKelurahan.ajax.reload();
+                                $('#dataKriteriaWilayah').DataTable().ajax.reload();
+                                tableKelurahan.page(currentPage).draw('page');
+                                Swal.fire({
+                                    title: 'Success',
+                                    text: response.message,
+                                    icon: 'success',
+                                });
+                            },
+                            error: function(xhr) {
+                                Swal.fire({
+                                    title: 'Error',
+                                    text: xhr.responseJSON.message,
+                                    icon: 'error',
+                                });
+                            }
+                        })
                     }
-                })
+                });
             });
         });
 
@@ -86,10 +107,18 @@
                     $('#dataKelurahan').DataTable().page(currentPage).draw('page');
                     $('#dataKriteriaWilayah').DataTable().ajax.reload();
                     $('#wilayahKelurahanModal').modal('hide');
-                    console.log(response.message);
+                    Swal.fire({
+                        title: 'Success',
+                        text: response.message,
+                        icon: 'success',
+                    });
                 },
                 error: function(xhr) {
-                    console.log(xhr.responseJSON.message);
+                    Swal.fire({
+                        title: 'Error',
+                        text: xhr.responseJSON.message,
+                        icon: 'error',
+                    });
                 }
             });
         });
